requirements.txt
# Core dependencies for local transcription and diarization
faster-whisper
pyannote.audio
# For diarization, you may need a HuggingFace token for pyannote.audio models. 